A few adjustments Here
We updated the UI transforms for easier quality of life basically made it so the childs axis doesn't change when editing in imgui.

Added some color for the debug cube to represent some depth.

Cleaned up the gameLogic: Deleted the commented out code

Sight change to the vcpkg setup. Mainly install vcpkg in your user profile so you need to make one less change when setting up. There is still a change you'll need to make the buit in baseline